requirements.txt
torch>=2.0.0
transformers>=4.31.0
bitsandbytes>=0.39.0
accelerate>=0.21.0
datasets>=2.12.0
pandas>=2.0.3
python-dotenv>=1.0.0
